moved differsFromLocalNameFormat to /queries/preference
