+++ app/actions/websocket/preferences.ts
@@ -5,9 +5,8 @@ import {updateDmGmDisplayName} from '@actions/local/channel';
 import {fetchPostById} from '@actions/remote/post';
 import {Preferences} from '@constants';
 import DatabaseManager from '@database/manager';
-import {getPreferenceValue} from '@helpers/api/preference';
 import {getPostById} from '@queries/servers/post';
-import {deletePreferences, queryPreferencesByCategoryAndName} from '@queries/servers/preference';
+import {deletePreferences, differsFromLocalNameFormat} from '@queries/servers/preference';
 
 export async function handlePreferenceChangedEvent(serverUrl: string, msg: WebSocketMessage): Promise<void> {
     const operator = DatabaseManager.serverDatabases[serverUrl].operator;
@@ -92,21 +91,3 @@ async function handleSavePostAdded(serverUrl: string, preferences: PreferenceTyp
     }
 }
 
-const differsFromLocalNameFormat = async (serverUrl: string, preferences: PreferenceType[]) => {
-    const database = DatabaseManager.serverDatabases[serverUrl]?.database;
-    if (!database) {
-        return false;
-    }
-
-    const displayPref = getPreferenceValue(preferences, Preferences.CATEGORY_DISPLAY_SETTINGS, Preferences.NAME_NAME_FORMAT) as string;
-    if (displayPref === '') {
-        return false;
-    }
-
-    const currentPref = await queryPreferencesByCategoryAndName(database, Preferences.CATEGORY_DISPLAY_SETTINGS, Preferences.NAME_NAME_FORMAT, displayPref).fetch();
-    if (currentPref.length > 0) {
-        return false;
-    }
-
-    return true;
-};
